requirements.txt
Django==3.0.5
python-decouple==3.3
djangorestframework==3.11.0
requests==2.22.0
short-url==1.2.2
django-widget-tweaks==1.4.8
